% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/012-roc.R
\name{auc.roc}
\alias{auc.roc}
\title{Compute ROC curves and AUC for each evaluation column.}
\usage{
\method{auc}{roc}(x, cols = NULL, ...)
}
\arguments{
\item{x}{roc object}

\item{cols}{column names to analyze; default is all}

\item{...}{reserved arguments}
}
\value{
Updated roc object with results stored in $auc_list and $auc_table
}
\description{
Compute ROC curves and AUC for each evaluation column.
}
\examples{
  obj <- roc(ivd_roc_example, cols = c("x1", "x2"), reference = "ref")
  obj <- auc(obj)
  obj <- auc(obj, cols = "x1")
}
