% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-generics.R
\name{describe}
\alias{describe}
\alias{correlation}
\alias{regression}
\alias{bias}
\alias{bland_altman}
\alias{report}
\alias{profile}
\alias{normal}
\alias{ci}
\alias{variance}
\alias{vc}
\alias{diagnostics}
\alias{kappa}
\alias{mcnemar}
\alias{cutoff}
\alias{mlr}
\alias{auc}
\alias{auc_compare}
\alias{tukey}
\title{Describe an analysis object}
\usage{
describe(x, ...)

correlation(x, ...)

regression(x, ...)

bias(x, ...)

bland_altman(x, ...)

report(x, ...)

profile(object, ...)

normal(x, ...)

ci(x, ...)

variance(x, ...)

vc(x, ...)

diagnostics(object, ...)

kappa(x, ...)

mcnemar(x, ...)

cutoff(x, ...)

mlr(x, ...)

auc(x, ...)

auc_compare(x, ...)

tukey(x, ...)
}
\arguments{
\item{x}{An S3 analysis object.}

\item{...}{Additional arguments passed to a method.}

\item{object}{An S3 analysis object.}
}
\value{
The value returned by the dispatched method.

The value returned by the dispatched method.

The value returned by the dispatched method.

The value returned by the dispatched method.

The value returned by the dispatched method.

The value returned by the dispatched method.

The value returned by the dispatched method.

The value returned by the dispatched method.

The value returned by the dispatched method.

The value returned by the dispatched method.

The value returned by the dispatched method.

The value returned by the dispatched method.

The value returned by the dispatched method.

The value returned by the dispatched method.

The value returned by the dispatched method.

The value returned by the dispatched method.

The value returned by the dispatched method.

The value returned by the dispatched method.

The value returned by the dispatched method.
}
\description{
Describe an analysis object
}
\examples{
obj <- mcr(ivd_mcr_example, "sid", "test", "ref")
describe(obj)
}
