% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/012-roc.R
\name{describe.roc}
\alias{describe.roc}
\title{Descriptive Statistics (S3 method)}
\usage{
\method{describe}{roc}(x, cols = NULL, digits = 4L, ...)
}
\arguments{
\item{x}{roc object}

\item{cols}{columns to describe; defaults to all evaluation columns + reference column (if numeric)}

\item{digits}{number of decimal places, default 4}

\item{...}{reserved arguments}
}
\value{
Updated roc object with results stored in $describe
}
\description{
Print distribution summaries for evaluation columns and the reference column (if numeric).
When the reference is binary, evaluation column distributions are shown grouped by positive/negative.
}
\examples{
  obj <- roc(ivd_roc_example, cols = "x1", reference = "ref")
  describe(obj)
}
