% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/003-example-data.R
\docType{data}
\name{ivd_bottle_example}
\alias{ivd_bottle_example}
\title{Small deterministic one-way ANOVA example}
\format{
A data frame with 12 observations from three bottles.
}
\usage{
ivd_bottle_example
}
\description{
Small deterministic one-way ANOVA example
}
\keyword{datasets}
