% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/003-example-data.R
\docType{data}
\name{ivd_mcr_example}
\alias{ivd_mcr_example}
\title{Small deterministic method-comparison example}
\format{
A data frame with 12 rows and columns \code{sid}, \code{test}, and \code{ref}.
}
\usage{
ivd_mcr_example
}
\description{
Small deterministic method-comparison example
}
\keyword{datasets}
