% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/003-example-data.R
\docType{data}
\name{ivd_qualitative_example}
\alias{ivd_qualitative_example}
\title{Small deterministic qualitative-method example}
\format{
A data frame with paired binary results and sample identifiers.
}
\usage{
ivd_qualitative_example
}
\description{
Small deterministic qualitative-method example
}
\keyword{datasets}
