% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/002-ivdtools-package.R
\docType{package}
\name{ivdtools-package}
\alias{ivdtools}
\alias{ivdtools-package}
\title{ivdtools: Statistical Tools for In Vitro Diagnostic Method Evaluation}
\description{
The package provides a collection of statistical workflows for evaluating
in vitro diagnostic methods and reagents. Most multi-step workflows use S3
objects: construct an analysis object, add results with analysis generics,
and inspect the accumulated results with \code{summary()} or \code{plot()}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hiox-tech/ivdtools}
  \item Report bugs at \url{https://github.com/hiox-tech/ivdtools/issues}
}

}
\author{
\strong{Maintainer}: hiox-tech \email{GeorgeBinDragon@outlook.com} [copyright holder]

}
\keyword{internal}
