% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-generics.R
\name{outlier}
\alias{outlier}
\title{Detect outliers in an analysis object}
\usage{
outlier(x, ...)
}
\arguments{
\item{x}{An S3 analysis object.}

\item{...}{Additional arguments passed to a method.}
}
\value{
The value returned by the dispatched method.
}
\description{
Detect outliers in an analysis object
}
\examples{
  obj <- mcr(ivd_mcr_example, "sid", "test", "ref")
  outlier(obj, method = "grubbs")
}
