% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/021-commutability.R
\name{plot.commutability_result}
\alias{plot.commutability_result}
\title{Plot a commutability assessment}
\usage{
\method{plot}{commutability_result}(x, method = NULL, ...)
}
\arguments{
\item{x}{A \code{commutability_result} object.}

\item{method}{For EP30, either \code{"prediction"} or \code{"relative_residual"}.
Ignored by the other approaches.}

\item{...}{Reserved arguments.}
}
\value{
A ggplot object, invisibly.
}
\description{
Plot a commutability assessment
}
\examples{
clinical <- expand.grid(
  sample = paste0("CS", 1:8), procedure = c("MP1", "MP2"),
  replicate = 1:3, stringsAsFactors = FALSE
)
level <- setNames(seq(10, 80, length.out = 8), paste0("CS", 1:8))
clinical$result <- level[clinical$sample] +
  ifelse(clinical$procedure == "MP2", 1, 0) +
  rep(c(-0.2, 0, 0.2), each = 16)
clinical$material_type <- "clinical"
clinical$position <- NA_integer_
rm <- expand.grid(
  sample = "RM1", procedure = c("MP1", "MP2"),
  position = 1:3, replicate = 1:3, stringsAsFactors = FALSE
)
rm$result <- 45 + ifelse(rm$procedure == "MP2", 1, 0) +
  rep(c(-0.1, 0, 0.1), each = 6)
rm$material_type <- "rm"
columns <- c("sample", "material_type", "procedure", "result",
             "position", "replicate")
example_data <- rbind(clinical[columns], rm[columns])
assessment <- commutability(
  example_data, "sample", "material_type", "procedure", "result"
)
plot(assessment)
}
