% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/017-interference.R
\name{plot.interference_dose_response}
\alias{plot.interference_dose_response}
\title{Plot interference dose-response results}
\usage{
\method{plot}{interference_dose_response}(x, type = c("response", "absolute", "percent"), ...)
}
\arguments{
\item{x}{An \code{interference_dose_response} object.}

\item{type}{\code{"response"}, \code{"absolute"}, or \code{"percent"}.}

\item{...}{Unused.}
}
\value{
A ggplot object.
}
\description{
Plot interference dose-response results
}
\examples{
d <- expand.grid(conc = c(0, 10, 20), rep = 1:5)
d$result <- 100 + 0.5 * d$conc
plot(interference_dose_response(d, "result", "conc"), type = "percent")
}
