% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/017-interference.R
\name{plot.interference_paired}
\alias{plot.interference_paired}
\title{Plot paired interference estimates}
\usage{
\method{plot}{interference_paired}(x, ...)
}
\arguments{
\item{x}{An \code{interference_paired} object.}

\item{...}{Unused.}
}
\value{
A ggplot object.
}
\description{
Plot paired interference estimates
}
\examples{
d <- data.frame(group = rep(c("control", "test"), each = 5),
                result = c(10, 10.1, 9.9, 10.2, 9.8, 11, 11.1, 10.9, 11.2, 10.8))
plot(interference_paired(d, "result", "ep07", condition = "group"))
}
