% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/017-interference.R
\name{plot.interference_patient}
\alias{plot.interference_patient}
\title{Plot patient-specimen interference results}
\usage{
\method{plot}{interference_patient}(x, type = c("difference", "interferent"), interferent = NULL, ...)
}
\arguments{
\item{x}{An \code{interference_patient} object.}

\item{type}{\code{"difference"} or \code{"interferent"}.}

\item{interferent}{Interferent concentration column to plot when
\code{type = "interferent"}.}

\item{...}{Unused.}
}
\value{
A ggplot object.
}
\description{
Plot patient-specimen interference results
}
\examples{
d <- expand.grid(id = paste0("S", 1:8), method = c("test", "ref"), rep = 1:2)
d$group <- ifelse(as.integer(sub("S", "", d$id)) <= 4, "control", "selected")
d$bilirubin <- 2 * (as.integer(sub("S", "", d$id)) - 1)
d$result <- 20 + as.integer(sub("S", "", d$id)) +
  ifelse(d$method == "test", 0.1 * d$bilirubin, 0)
fit <- interference_patient(d, "id", "group", "method", "result",
  "selected", "control", "test", "ref", "bilirubin")
plot(fit, type = "difference")
}
