% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/015-sensitivity.R
\name{plot.sensitivity}
\alias{plot.sensitivity}
\title{Plot an analytical-sensitivity analysis}
\usage{
\method{plot}{sensitivity}(x, ...)
}
\arguments{
\item{x}{A sensitivity-analysis result object.}

\item{...}{Reserved for the S3 generic.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot an analytical-sensitivity analysis
}
\examples{
d <- data.frame(sample = rep("blank", 10), result = seq(0.1, 1, length.out = 10))
fit <- sensitivity_lob(d, "result", "sample")
plot(fit)
}
