% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/010-qualitative-analysis.R
\name{print.describe_table}
\alias{print.describe_table}
\title{S3 print method for describe_table}
\usage{
\method{print}{describe_table}(x, ...)
}
\arguments{
\item{x}{describe_table object}

\item{...}{reserved arguments}
}
\value{
Invisible describe_table object
}
\description{
S3 print method for describe_table
}
\examples{
  df <- data.frame(
    new    = c("positive","positive","negative","negative","positive","negative"),
    gold   = c("positive","negative","positive","negative","positive","positive"),
    age    = c(45, 52, 38, 61, 47, 55),
    gender = c("M","F","F","M","M","F"),
    id     = c("S001","S002","S003","S004","S005","S006")
  )
  tab <- raw_to_table(df, "new", "gold", id = "id")
  tab <- describe(tab)
}
