% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/007-method-comparison.R
\name{print.factor_split}
\alias{print.factor_split}
\title{Print a categorical data split}
\usage{
\method{print}{factor_split}(x, ...)
}
\arguments{
\item{x}{A \code{factor_split} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
The input object, invisibly.
}
\description{
Print a categorical data split
}
