% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005-fit.R
\name{print.fit_equation}
\alias{print.fit_equation}
\title{Print an equation fit}
\usage{
\method{print}{fit_equation}(x, ...)
}
\arguments{
\item{x}{A \code{fit_equation} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
\code{x}, invisibly.
}
\description{
Print an equation fit
}
