% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/017-interference.R
\name{print.interference_dose_response}
\alias{print.interference_dose_response}
\title{Print interference dose-response results}
\usage{
\method{print}{interference_dose_response}(x, ...)
}
\arguments{
\item{x}{An \code{interference_dose_response} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print interference dose-response results
}
\examples{
d <- expand.grid(conc = c(0, 10, 20), rep = 1:5)
d$result <- 100 + 0.5 * d$conc
print(interference_dose_response(d, "result", "conc"))
}
