% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/017-interference.R
\name{print.interference_patient}
\alias{print.interference_patient}
\title{Print patient-specimen interference results}
\usage{
\method{print}{interference_patient}(x, ...)
}
\arguments{
\item{x}{An \code{interference_patient} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print patient-specimen interference results
}
\examples{
d <- expand.grid(id = paste0("S", 1:6), method = c("test", "ref"), rep = 1:2)
d$group <- ifelse(as.integer(sub("S", "", d$id)) <= 3, "control", "selected")
d$result <- 10 + as.integer(sub("S", "", d$id)) + (d$method == "test")
print(interference_patient(d, "id", "group", "method", "result",
  "selected", "control", "test", "ref"))
}
