% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/017-interference.R
\name{print.interference_replicates}
\alias{print.interference_replicates}
\title{Print EP07 replicate-count results}
\usage{
\method{print}{interference_replicates}(x, ...)
}
\arguments{
\item{x}{An \code{interference_replicates} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print EP07 replicate-count results
}
\examples{
print(interference_replicates(5, 7))
}
