% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/016-linearity.R
\name{print.linearity}
\alias{print.linearity}
\title{Print a linearity analysis}
\usage{
\method{print}{linearity}(x, ...)
}
\arguments{
\item{x}{A \code{linearity} object.}

\item{...}{Reserved arguments.}
}
\value{
The unchanged \code{x}, invisibly.
}
\description{
Print a linearity analysis
}
