% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/007-method-comparison.R
\name{print.mcr}
\alias{print.mcr}
\title{S3 print method}
\usage{
\method{print}{mcr}(x, ...)
}
\arguments{
\item{x}{\code{mcr} object}

\item{...}{Additional arguments}
}
\value{
Invisible \code{mcr} object
}
\description{
Print data overview and analysis slot status based on \code{$print} slot.
}
\examples{
  df <- data.frame(sid = 1:30,
                   test = rnorm(30, 50, 10),
                   ref  = rnorm(30, 50, 10))
  obj <- mcr(df, "sid", "test", "ref")
  print(obj)
}
