% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/007-method-comparison.R
\name{print.mcr_bias}
\alias{print.mcr_bias}
\title{Print method for bias results}
\usage{
\method{print}{mcr_bias}(x, ...)
}
\arguments{
\item{x}{mcr_bias object}

\item{...}{additional arguments}
}
\value{
The unchanged \code{mcr_bias} data frame, invisibly.
}
\description{
Print method for bias results
}
