% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/007-method-comparison.R
\name{print.mcr_bland_altman}
\alias{print.mcr_bland_altman}
\title{Print method for Bland-Altman results}
\usage{
\method{print}{mcr_bland_altman}(x, ...)
}
\arguments{
\item{x}{mcr_bland_altman object}

\item{...}{additional arguments}
}
\value{
The unchanged \code{mcr_bland_altman} object, invisibly.
}
\description{
Print method for Bland-Altman results
}
