% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/007-method-comparison.R
\name{print.mcr_correlation}
\alias{print.mcr_correlation}
\title{Print method for correlation results}
\usage{
\method{print}{mcr_correlation}(x, ...)
}
\arguments{
\item{x}{mcr_correlation object}

\item{...}{additional arguments}
}
\value{
The unchanged \code{mcr_correlation} object, invisibly.
}
\description{
Print method for correlation results
}
