% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/007-method-comparison.R
\name{print.mcr_describe}
\alias{print.mcr_describe}
\title{Print method for describe results}
\usage{
\method{print}{mcr_describe}(x, digits = NULL, ...)
}
\arguments{
\item{x}{mcr_describe object}

\item{digits}{Number of decimal places, default 4}

\item{...}{additional arguments}
}
\value{
The unchanged \code{mcr_describe} object, invisibly.
}
\description{
Print method for describe results
}
