% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/007-method-comparison.R
\name{print.mcr_outlier}
\alias{print.mcr_outlier}
\title{Print method for outlier results}
\usage{
\method{print}{mcr_outlier}(x, ...)
}
\arguments{
\item{x}{mcr_outlier object}

\item{...}{additional arguments}
}
\value{
The unchanged \code{mcr_outlier} object, invisibly.
}
\description{
Print method for outlier results
}
