% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/007-method-comparison.R
\name{print.mcr_regression}
\alias{print.mcr_regression}
\title{Print method for regression results}
\usage{
\method{print}{mcr_regression}(x, ...)
}
\arguments{
\item{x}{mcr_regression object}

\item{...}{additional arguments}
}
\value{
The unchanged \code{mcr_regression} object, invisibly.
}
\description{
Print method for regression results
}
