% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/008-precision.R
\name{print.precision}
\alias{print.precision}
\title{Print a precision object overview}
\usage{
\method{print}{precision}(x, ...)
}
\arguments{
\item{x}{\code{precision} object}

\item{...}{Reserved arguments}
}
\value{
Invisible \code{precision} object
}
\description{
Print the data overview of a \code{precision} object (data class, total rows, formula,
grouping variables, sample count, factor levels, balance warnings) and the analysis
slot status (\code{[x]} / \code{[ ]}).
}
