% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/008-precision.R
\name{print.precision_normal}
\alias{print.precision_normal}
\title{Print normality test results}
\usage{
\method{print}{precision_normal}(x, ...)
}
\arguments{
\item{x}{\code{precision_normal} object}

\item{...}{Reserved arguments}
}
\value{
Invisibly returns \code{x}.
}
\description{
Prints the selected normality test result per sample, including the actual
method, sample size, statistic name, statistic, and p-value.
}
