% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/008-precision.R
\name{print.precision_vc}
\alias{print.precision_vc}
\title{Print variance component table}
\usage{
\method{print}{precision_vc}(x, ...)
}
\arguments{
\item{x}{\code{precision_vc} object (a data frame)}

\item{...}{Reserved arguments}
}
\value{
Invisibly returns \code{x}.
}
\description{
Prints the variance component summary table with columns: sample, component,
VC (variance component), \\%Total, SD, and CV\link[=\\\%]{\\%}.
}
