% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/011-reference-interval.R
\name{print.reference_interval}
\alias{print.reference_interval}
\title{Print reference interval analysis results}
\usage{
\method{print}{reference_interval}(x, ...)
}
\arguments{
\item{x}{A "reference_interval" object.}

\item{...}{Additional arguments (reserved for generic).}
}
\value{
The unchanged \code{reference_interval} object, invisibly.
}
\description{
Print reference interval analysis results
}
