% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/012-roc.R
\name{print.roc}
\alias{print.roc}
\title{Print a ROC analysis object}
\usage{
\method{print}{roc}(x, ...)
}
\arguments{
\item{x}{A \code{roc} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
\code{x}, invisibly.
}
\description{
Print a ROC analysis object
}
