% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/012-roc.R
\name{print.roc_auc_comparison}
\alias{print.roc_auc_comparison}
\title{Print a paired AUC comparison}
\usage{
\method{print}{roc_auc_comparison}(x, ...)
}
\arguments{
\item{x}{A \code{roc_auc_comparison} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
The input object, invisibly.
}
\description{
Print a paired AUC comparison
}
