% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/012-roc.R
\name{print.roc_auc_table}
\alias{print.roc_auc_table}
\title{Print method for roc AUC table}
\usage{
\method{print}{roc_auc_table}(x, ...)
}
\arguments{
\item{x}{roc_auc_table object (data.frame)}

\item{...}{additional arguments}
}
\value{
The unchanged \code{roc_auc_table} data frame, invisibly.
}
\description{
Print method for roc AUC table
}
