% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/012-roc.R
\name{print.roc_cutoff_table}
\alias{print.roc_cutoff_table}
\title{Print method for roc cutoff table}
\usage{
\method{print}{roc_cutoff_table}(x, ...)
}
\arguments{
\item{x}{roc_cutoff_table object (data.frame)}

\item{...}{additional arguments}
}
\value{
The unchanged \code{roc_cutoff_table} data frame, invisibly.
}
\description{
Print method for roc cutoff table
}
