% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/012-roc.R
\name{print.roc_describe}
\alias{print.roc_describe}
\title{Print method for roc describe results}
\usage{
\method{print}{roc_describe}(x, digits = NULL, ...)
}
\arguments{
\item{x}{roc_describe object}

\item{digits}{number of decimal places, default 4}

\item{...}{additional arguments}
}
\value{
The unchanged \code{roc_describe} object, invisibly.
}
\description{
Print method for roc describe results
}
