% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/012-roc.R
\name{print.roc_mlr}
\alias{print.roc_mlr}
\title{Print a multivariate ROC model}
\usage{
\method{print}{roc_mlr}(x, ...)
}
\arguments{
\item{x}{A \code{roc_mlr} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
\code{x}, invisibly.
}
\description{
Print a multivariate ROC model
}
