% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/020-uncertainty.R
\name{print.uncertainty_characterization}
\alias{print.uncertainty_characterization}
\title{Print characterization uncertainty results}
\usage{
\method{print}{uncertainty_characterization}(x, ...)
}
\arguments{
\item{x}{An uncertainty_characterization object.}

\item{...}{Reserved for the S3 generic.}
}
\value{
The unchanged x, invisibly.
}
\description{
Print characterization uncertainty results
}
