% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/020-uncertainty.R
\name{print.uncertainty_homogeneity}
\alias{print.uncertainty_homogeneity}
\title{Print between-unit homogeneity uncertainty results}
\usage{
\method{print}{uncertainty_homogeneity}(x, ...)
}
\arguments{
\item{x}{An uncertainty_homogeneity object.}

\item{...}{Reserved for the S3 generic.}
}
\value{
The unchanged x, invisibly.
}
\description{
Print between-unit homogeneity uncertainty results
}
