% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/020-uncertainty.R
\name{print.uncertainty_iqc}
\alias{print.uncertainty_iqc}
\title{Print IQC uncertainty results}
\usage{
\method{print}{uncertainty_iqc}(x, ...)
}
\arguments{
\item{x}{An uncertainty_iqc object.}

\item{...}{Reserved for the S3 generic.}
}
\value{
The unchanged x, invisibly.
}
\description{
Print IQC uncertainty results
}
