% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/008-precision.R
\name{profile.precision}
\alias{profile.precision}
\title{S3 precision profile method}
\usage{
\method{profile}{precision}(object, model.no = 1:10, ...)
}
\arguments{
\item{object}{\code{precision} object (requires \code{variance()} to be run first)}

\item{model.no}{Vector of candidate model numbers, default \code{1:10}}

\item{...}{Additional arguments passed to \code{.sadler()} (e.g. \code{K})}
}
\value{
Updated \code{precision} object with results stored in \code{$profile}
}
\description{
Fit precision (SD) vs. mean across samples using the Sadler model selection algorithm
(10 candidate models) based on \code{VFP::fit_vfp()}.
}
\examples{
  data(VCAdata1, package = "VCA")
  obj <- precision(VCAdata1, y ~ day/run, by = "sample")
  obj <- variance(obj)
  obj <- profile(obj, model.no = 1)
}
