% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/010-qualitative-analysis.R
\name{summary.fourfold_table}
\alias{summary.fourfold_table}
\title{S3 summary method}
\usage{
\method{summary}{fourfold_table}(object, digits = 3, ...)
}
\arguments{
\item{object}{a fourfold_table object}

\item{digits}{Number of decimal places, default 3}

\item{...}{Reserved arguments}
}
\value{
Invisible fourfold_table object
}
\description{
Summarizes all key analysis information already executed in a fourfold_table object,
formatted following summary.mcr() pattern (mcr.R).
}
\examples{
  tab <- counts_to_table(tp = 85, fp = 3, tn = 90, fn = 2,
                         candidate = "new method", reference = "gold standard")
  summary(tab)
  tab <- diagnostics(tab); tab <- kappa(tab); tab <- mcnemar(tab)
  summary(tab)
}
