% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/007-method-comparison.R
\name{summary.mcr}
\alias{summary.mcr}
\title{S3 summary method}
\usage{
\method{summary}{mcr}(object, ...)
}
\arguments{
\item{object}{\code{mcr} object}

\item{...}{Additional arguments}
}
\value{
Invisible \code{mcr} object
}
\description{
Summarize key information from all executed analyses in the \code{mcr} object
(descriptive statistics, correlation analysis, regression analysis,
outlier detection, Bland-Altman analysis, and bias analysis).
}
\examples{
  df <- data.frame(sid = 1:30,
                   test = rnorm(30, 50, 10),
                   ref  = rnorm(30, 50, 10))
  obj <- mcr(df, "sid", "test", "ref")
  summary(obj)
  obj <- correlation(obj)
  obj <- regression(obj, method = "ols")
  obj <- bland_altman(obj)
  summary(obj)
}
