% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/008-precision.R
\name{summary.precision}
\alias{summary.precision}
\title{S3 summary method}
\usage{
\method{summary}{precision}(object, ...)
}
\arguments{
\item{object}{\code{precision} object}

\item{...}{Reserved arguments}
}
\value{
Invisible \code{precision} object
}
\description{
Summarize all executed analyses in a \code{precision} object, including data description,
analysis status checklist, outlier detection, normality tests, variance components,
confidence intervals, and precision profile.
}
\examples{
  data(VCAdata1, package = "VCA")
  obj <- precision(VCAdata1, y ~ day/run, by = "sample")
  summary(obj)
}
