% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/012-roc.R
\name{summary.roc}
\alias{summary.roc}
\title{Summary (S3 method)}
\usage{
\method{summary}{roc}(object, ...)
}
\arguments{
\item{object}{roc object}

\item{...}{reserved arguments}
}
\value{
invisible roc object
}
\description{
Output key information from all analyses performed on the roc object.
}
\examples{
  obj <- roc(ivd_roc_example, cols = c("x1", "x2"), reference = "ref")
  summary(obj)
  obj <- auc(obj)
  obj <- cutoff(obj)
  obj <- mlr(obj)
  summary(obj)
}
